separate page component to display different contents for when signed-in and signed-out
