+++ components/SignedInContent.tsx
@@ -1,0 +1,10 @@
+import NavbarDemo from "@/components/Header";
+
+export default function SignedInContent() {
+  return (
+    <>
+      <NavbarDemo />
+      <h1>Welcome to your Dashboard</h1>
+    </>
+  );
+}
